// src/app/auth/types/authState.interface.ts
import { backEndErrorsInterface } from "src/app/shared/types/backEndErrors.interface"
import { CurrentUserInterface } from "src/app/shared/types/currentUser.interface"

export interface AuthStateInterface {
  isSubmitting: boolean
  isLoading: boolean
  currentUser: CurrentUserInterface | null
  isLoggedIn: boolean | null
  validationErrors: backEndErrorsInterface | null
}